requirement.txt
certifi==2021.10.8
charset-normalizer==2.0.12
future==0.18.2
idna==3.3
oauthlib==3.2.0
pypassage==1.3.1
python-scriptures==3.0.0
requests==2.27.1
requests-oauthlib==1.3.1
tweepy==4.5.0
urllib3==1.26.8
